requirements.txt.txt
flask
tensorflow-cpu
pillow
pydicom
opencv-python-headless